+++ iris_heap_lang/proofmode.v
@@ -1,4 +1,4 @@
-From iris.proofmode Require Import coq_tactics reduction.
+From iris.proofmode Require Import coq_tactics reduction spec_patterns.
 From iris.proofmode Require Export tactics.
 From iris.program_logic Require Import atomic.
 From iris.heap_lang Require Export tactics derived_laws.
@@ -556,9 +556,13 @@ Tactic Notation "awp_apply" open_constr(lem) :=
   wp_apply_core lem ltac:(fun H => iApplyHyp H) ltac:(fun cont => fail);
   last iAuIntro.
 Tactic Notation "awp_apply" open_constr(lem) "without" constr(Hs) :=
+  (* Convert "list of hypothesis" into specialization pattern. *)
+  let Hs := words Hs in
+  let Hs := eval vm_compute in (INamed <$> Hs) in
   wp_apply_core lem
     ltac:(fun H =>
-      iApply wp_frame_wand_l; iSplitL Hs; [iAccu|iApplyHyp H]) 
+      iApply (wp_frame_wand with
+        [SGoal $ SpecGoal GSpatial false [] Hs false]); [iAccu|iApplyHyp H])
     ltac:(fun cont => fail);
   last iAuIntro.